+++ macro/dumpEvent.py
@@ -3,11 +3,11 @@
 
 # example for dumping an MC event
 
+import geometry_config
 import ROOT
-import ShipGeoConfig
 import shipunit as u
 
-ship_geo = ShipGeoConfig.Config().loadpy("$FAIRSHIP/geometry/geometry_config.py")
+ship_geo = geometry_config.create_config(shieldName="TRY_2025")
 PDG = ROOT.TDatabasePDG.Instance()